+++ pydantic_resolve/resolver.py
@@ -7,21 +7,27 @@ from .exceptions import ResolverTargetAttrNotFound, LoaderFieldNotProvidedError,
 from typing import Any, Callable, Optional
 from pydantic_resolve import core
 from .constant import PREFIX
-from .util import get_class_field_annotations
+from .util import get_class_field_annotations, replace_method, transformer_decorator
 from inspect import isclass
 from aiodataloader import DataLoader
 
 
 def LoaderDepend(  # noqa: N802
-    dependency: Optional[Callable[..., Any]] = None 
+    dependency: Optional[Callable[..., Any]] = None,
+    transform: Optional[Callable[..., Any]] = None 
 ) -> Any:
-    return Depends(dependency=dependency)
+    return Depends(
+        dependency=dependency,
+        transform=transform)
 
 class Depends:
     def __init__(
-        self, dependency: Optional[Callable[..., Any]] = None
+        self, 
+        dependency: Optional[Callable[..., Any]] = None,
+        transform: Optional[Callable[..., Any]] = None
     ):
         self.dependency = dependency
+        self.transform = transform
 
 T = TypeVar("T")
 
@@ -41,28 +47,49 @@ class Resolver:
         # manage the creation of loader instances
         for k, v in signature.parameters.items():
             if isinstance(v.default, Depends):
+                # Base: DataLoader or batch_load_fn
+                # transform: data transform function
+                Base = v.default.dependency
+                transform = v.default.transform
+
                 # module.kls to avoid same kls name from different module
                 cache_key = f'{v.default.dependency.__module__}.{v.default.dependency.__name__}'
-                cache_provider = self.ctx.get()
+                if transform:
+                    # lambda's name is "lambda", not distinctive.
+                    cache_key = f'{cache_key}:{transform.__module__}.{id(transform)}'  
 
+                cache_provider = self.ctx.get()
                 hit = cache_provider.get(cache_key, None)
+
                 if hit:
                     loader = hit
                 else:
                     # create loader instance 
-                    if isclass(v.default.dependency):
-                        loader = v.default.dependency()
+                    if isclass(Base):
+                        # if extra transform provides
+                        if transform:
+
+                            # replace with new batch_fn with transform function
+                            async def batch_load_fn(self, *args):
+                                items = await Base.batch_load_fn(self, *args)  # type:ignore
+                                items = [transform(item) for item in items]  # type:ignore
+                                return items
+
+                            NewBase = replace_method(Base, 'DataLoader', 'batch_load_fn', batch_load_fn)
+                            loader = NewBase()
+                        else:
+                            loader = Base()
 
                         # and pick config from 'loader_filters' param, only for DataClass
                         filter_config_provider = self.loader_filters_ctx.get()
-                        filter_config = filter_config_provider.get(v.default.dependency, {})
+                        filter_config = filter_config_provider.get(Base, {})
 
                         # class ExampleLoader(DataLoader):
                         #     filtar_x: bool  <--------------- set this
                         #
                         #     async def batch_load_fn(self, keys):
                         #         ....
-                        for field in get_class_field_annotations(v.default.dependency):
+                        for field in get_class_field_annotations(Base):
                             try:
                                 value = filter_config[field]
                                 setattr(loader, field, value)
@@ -71,7 +98,11 @@ class Resolver:
 
                     # build loader from batch_load_fn, filters config is impossible
                     else:
-                        loader = DataLoader(batch_load_fn=v.default.dependency)
+                        if transform:
+                            _batch_load_fn = transformer_decorator(transform)(Base)  # type:ignore
+                        else:
+                            _batch_load_fn = Base
+                        loader = DataLoader(batch_load_fn=_batch_load_fn) # type:ignore
 
                     cache_provider[cache_key] = loader
                     self.ctx.set(cache_provider)